requirement.txt
requests
pandas
psycopg2-binary
datetime
selenium #futuramente